consolidate styles, promote more overrides
