+++ src/components/Addons/NewAddon.jsx
@@ -1,28 +1,11 @@
 import React, { Component } from 'react';
-import deepmerge from 'deepmerge';
 import PropTypes from 'prop-types';
 import {
   Step, Stepper, StepLabel, CircularProgress, LinearProgress, Button, Typography,
 } from '@material-ui/core';
-import { MuiThemeProvider } from '@material-ui/core/styles';
-import Select from '../Select';
+import Search from '../Search';
 import api from '../../services/api';
 import ConfirmationModal from '../ConfirmationModal';
-
-const theme = parentTheme => deepmerge(parentTheme, {
-  overrides: {
-    MuiButton: {
-      root: {
-        marginRight: '15px',
-      },
-    },
-    MuiFormControl: {
-      root: {
-        marginBottom: '15px',
-      },
-    },
-  },
-});
 
 const style = {
   stepper: {
@@ -202,7 +185,7 @@ export default class NewAddon extends Component {
         return (
           <div>
             <div style={style.selectContainer}>
-              <Select
+              <Search
                 options={groupedServices}
                 value={service}
                 onChange={this.handleServiceChange}
@@ -218,7 +201,7 @@ export default class NewAddon extends Component {
         return (
           <div>
             <div style={style.selectContainer}>
-              <Select
+              <Search
                 options={plans}
                 value={plan ? plan.id : ''}
                 onChange={this.handlePlanChange}
@@ -273,6 +256,7 @@ export default class NewAddon extends Component {
               className="back"
               disabled={stepIndex === 0}
               onClick={this.handlePrev}
+              style={style.buttons.back}
             >
               Back
             </Button>
@@ -319,7 +303,7 @@ export default class NewAddon extends Component {
     const renderCaption = text => <Typography variant="caption" className="step-label-caption">{text}</Typography>;
 
     return (
-      <MuiThemeProvider theme={theme}>
+      <div>
         <div style={provisionStyle}>
           <Stepper activeStep={stepIndex}>
             <Step>
@@ -357,7 +341,7 @@ export default class NewAddon extends Component {
           <label>{provisionMessage}</label> {/* eslint-disable-line */}
           <LinearProgress variant="determinate" value={provisionStatus * 100} />
         </div>
-      </MuiThemeProvider>
+      </div>
     );
   }
 }
